% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompCCAT.R
\name{CompCCAT}
\alias{CompCCAT}
\alias{CCAT}
\title{Correlation of Connectome And Transcriptome}
\usage{
CompCCAT(exp.m, ppiA.m)
}
\arguments{
\item{exp.m}{A scRNA-Seq data matrix with rows labeling genes and columns 
labeling single cells and with rownames annotated to a gene-identifier
identical to that used in the \code{ppiA.m} argument. scRNA-Seq data matrix
should have undergone prior QC to remove poor quality cells and each cell
normalized by library size. If data has not been log-transformed, the
function will log2-transform with a pseudocount of +1.}

\item{ppiA.m}{The adjacency matrix of a user-given PPI network with rownames and 
colnames labeling genes (same gene identifier as in \code{exp.m}). Diagonal
entries should be zero and the number of genes in the network should be large,
i.e. at least over 8000, to ensure a reasonably large overlap of genes with
those in the expression data matrix.}
}
\value{
CCAT
The estimated CCAT values as a vector
}
\description{
This function leverages Pearson correlation between gene expresion level
and gene connectome derived from PPI network to fastly estimate signaling
entropy rate.
}
\references{
Teschendorff Andrew E., Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.
}
