% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffusionLiteralPlot.R
\name{diffPlot}
\alias{diffPlot}
\title{Plotting the Diffusion Model}
\usage{
diffPlot(v, z, t0, a, zr, sv, szr, st0, ymin = NULL, ymax = NULL,
  tmin = NULL, tmax = NULL, tExtFactor = 1.2, topRTs = NULL,
  bottomRTs = NULL, topLab = NULL, bottomLab = NULL, aLty = 2,
  aLwd = 1, aCol = "gray", vLty = 1, vLwd = 2, vCol = "blue",
  vLength = 0.2, vAngle = 20, vCode = 2, t0Lty = 1, t0Lwd = 1,
  t0Col = "magenta", t0Length = 0.1, t0Angle = 90, t0Code = 3,
  xlab = "time (seconds)", ylab = "evidence", main = "",
  tLimit = 2.5, plotDensities = F, plotRtPts = F)
}
\arguments{
\item{v}{The drift rate.}

\item{z}{The starting point. Note that it is in the metric with a range 0 to
\code{a}. Either this or \code{zr} must be given.}

\item{t0}{Nondecision time.}

\item{a}{Threshold value.}

\item{zr}{Relative starting point. Used if \code{z} is missing, and should
represent a proportion of the \code{a} distance.}

\item{sv}{Variability of the drift rate.}

\item{szr}{Variability of the relative starting point.}

\item{st0}{Variability of nondecision time.}

\item{ymin}{Y (evidence) axis minimum value.}

\item{ymax}{Y (evidence) axis maximum value.}

\item{tmin}{Minimum time value to show.}

\item{tmax}{Maximum time value to show.}

\item{tExtFactor}{}

\item{topRTs}{}

\item{bottomRTs}{}

\item{topLab}{}

\item{bottomLab}{}

\item{aLty}{}

\item{aLwd}{}

\item{aCol}{}

\item{vLty}{}

\item{vLwd}{}

\item{vCol}{}

\item{vLength}{}

\item{vAngle}{}

\item{vCode}{}

\item{t0Lty}{}

\item{t0Lwd}{}

\item{t0Col}{}

\item{t0Length}{}

\item{t0Angle}{}

\item{t0Code}{}

\item{xlab}{}

\item{ylab}{}

\item{main}{}

\item{tLimit}{}
}
\description{
This function constructs a diffusion model plot. It shows average parameter
values.
}
