% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferPotencyStates.R
\name{InferPotencyStates}
\alias{InferPotencyStates}
\title{Infer distinct potency states from the single cell potency estimates}
\usage{
InferPotencyStates(
  potest.v,
  type = c("SR", "CCAT"),
  pheno.v = NULL,
  diffvar = TRUE,
  maxPS = 5
)
}
\arguments{
\item{potest.v}{A vector of potency estimates for all cells, e.g. a vector of SR or CCAT values.}

\item{type}{The type of potency estimate used (SR or CCAT).}

\item{pheno.v}{A phenotype vector for the single cells.}

\item{diffvar}{A logical. Default is TRUE.
Specifies whether the Gaussian mixture model to be fit assumes components 
to have different (default) or equal variance.
In the latter case, use *modelNames = c("E")*.}

\item{maxPS}{Maximum number of potency states, when inferring discrete potency 
states of single cells. Default value is 5.}
}
\value{
A list with the following elements:

potS
Inferred discrete potency states for each single cell. It is indexed so 
that the index increases as the signaling entropy of the state decreases

distr
If phenotype information provided, it will be a table giving the 
distribution of single-cells across potency states and phenotypes

prob
Table giving the probabilities of each potency state per phenotype value

het
The normalised Shannon Index of potency per phenotype value
}
\description{
This function infers the discrete potency states of single cells and 
its distribution across the single cell population.
}
\references{
Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.
}
