% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFFunctions.R
\name{plotCDF}
\alias{plotCDF}
\title{Title}
\usage{
plotCDF(..., realData = NULL, plotPars = NULL, linesPars = NULL)
}
\arguments{
\item{...}{Parameters for \code{plotCDFValues}}

\item{realData}{A data frame or with 2 colums: the first should be an 
\code{rt} column and the second should be a \code{response} column that 
has a 0 for each lower boundary response and a 1 for upper boundardy
responses.}

\item{plotPars}{A named list of arguments to pass to the \code{plot()}
call.}

\item{linesPars}{A named list of arguments to pass to the \code{lines()}
call that plots the \code{realData} cdf (if provided).}
}
\description{
Title
}
