% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDFFunctions.R
\name{plotCDFValues}
\alias{plotCDFValues}
\title{Title}
\usage{
plotCDFValues(..., pars = NULL, plotcdfLoc = "./plot-cdf",
  outFile = "tempCDF.lst", retainOutputFile = F)
}
\arguments{
\item{...}{Named parameters from the diffusion model, if not given as a 
list in \code{pars}}

\item{pars}{Diffusion model parameter set. Should be in a named list.}

\item{plotcdfLoc}{The location of the plotcdf program.}

\item{outFile}{The temporary file that will hold results of plotcdf.}
}
\value{
A set of values for the predicted CDF.
}
\description{
Title
}
\details{
The parameters that must be provided are \code{a}, \code{zr}, \code{v},
\code{t0}, \code{d}, \code{szr}, \code{sv}, and \code{st0}.
}
