% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoIntegPPI.R
\name{DoIntegPPI}
\alias{DoIntegPPI}
\title{Integration of gene expression matrix and PPI network}
\usage{
DoIntegPPI(exp.m, ppiA.m)
}
\arguments{
\item{exp.m}{The scRNA-Seq data matrix normalized for library size and log2-transformed
with a pseudocount of 1.1}

\item{ppiA.m}{The adjacency matrix of a user-given PPI network with rownames and 
colnames labeling genes (same gene identifier as in \code{exp.m})}
}
\value{
A list with two elements:

expMC
Reduced expression matrix with genes in the maximally connected subnetwork

adjMC
Adjacency matrix of the maximally connected subnetwork
}
\description{
This function finds the common genes between the scRNA-Seq data matrix 
and the genes present in the PPI network, and constructs the maximally 
connected subnetwork and associated expression matrix for the computation 
of signaling entropy.
}
\references{
Chen, Weiyan, et al.
\emph{Single-cell landscape in mammary epithelium reveals 
bipotent-like cells associated with breast cancer risk 
and outcome.}
Communications Biology 2 (2019): 306.
doi:\href{https://doi.org/10.1038/s42003-019-0554-8}{
10.1038/s42003-019-0554-8}. 

Teschendorff AE, Tariq Enver. 
\emph{Single-cell entropy for accurate estimation of differentiation 
potency from a cell’s transcriptome.}
Nature communications 8 (2017): 15599.
doi:\href{https://doi.org/10.1038/ncomms15599}{
10.1038/ncomms15599}.

Teschendorff AE, Banerji CR, Severini S, Kuehn R, Sollich P. 
\emph{Increased signaling entropy in cancer requires the scale-free 
property of protein interaction networks.}
Scientific reports 5 (2015): 9646.
doi:\href{https://doi.org/10.1038/srep09646}{
10.1038/srep09646}.

Banerji, Christopher RS, et al. 
\emph{Intra-tumour signalling entropy determines clinical outcome 
in breast and lung cancer.}
PLoS computational biology 11.3 (2015): e1004115.
doi:\href{https://doi.org/10.1371/journal.pcbi.1004115}{
10.1371/journal.pcbi.1004115}.

Teschendorff, Andrew E., Peter Sollich, and Reimer Kuehn.
\emph{Signalling entropy: A novel network-theoretical framework 
for systems analysis and interpretation of functional omic data.}
Methods 67.3 (2014): 282-293.
doi:\href{https://doi.org/10.1016/j.ymeth.2014.03.013}{
10.1016/j.ymeth.2014.03.013}.

Banerji, Christopher RS, et al. 
\emph{Cellular network entropy as the energy potential in 
Waddington's differentiation landscape.}
Scientific reports 3 (2013): 3039.
doi:\href{https://doi.org/10.1038/srep03039}{
10.1038/srep03039}.
}
